+++ tests/conftest.py
@@ -65,6 +65,11 @@ def bad_resource_parameters_in_is_path():
 
 # Schemas
 @pytest.fixture
+def super_schema():
+    return fixture_models.super_schema()
+
+
+@pytest.fixture
 def schema_test_model():
     return fixture_models.schema_test_model()
 
@@ -224,6 +229,11 @@ def request_body():
 
 
 @pytest.fixture
+def path_item_object():
+    return fixture_swagger.path_item_object()
+
+
+@pytest.fixture
 def components_schemas_object():
     return fixture_swagger.components_schemas_object()
 
@@ -271,3 +281,38 @@ def components_callback_object():
 @pytest.fixture
 def server_object():
     return fixture_swagger.server_object()
+
+
+@pytest.fixture
+def open_api_object():
+    return fixture_swagger.open_api_object()
+
+
+@pytest.fixture
+def paths_object():
+    return fixture_swagger.paths_object()
+
+
+@pytest.fixture
+def external_documentation_object():
+    return fixture_swagger.external_documentation_object()
+
+
+@pytest.fixture
+def security():
+    return fixture_swagger.security()
+
+
+@pytest.fixture
+def example_object():
+    return fixture_swagger.example_object()
+
+
+@pytest.fixture
+def tags_object():
+    return fixture_swagger.tags_object()
+
+
+@pytest.fixture
+def media_types():
+    return fixture_swagger.media_types()