+++ tests/test_schema.py
@@ -1,7 +1,9 @@
+import json
+
 import pytest
 from flask_restful_swagger_3.exceptions import SchemaAlreadyExist
 
-from flask_restful_swagger_3 import Schema
+from flask_restful_swagger_3 import Schema, ExampleEncoder, swagger
 
 
 class TestSchema:
@@ -357,3 +359,307 @@ class TestSchema:
             'name': 'test',
         }
         assert SchemaLoadOnlyAndFalse(**user) == user_result
+
+    def test_schema_valid_when_required_is_list(self):
+        class SchemaRequiredList(Schema):
+            type = 'object'
+            properties = {
+                'id': {
+                    'type': 'int'
+                },
+                'name': {
+                    'type': 'string',
+                },
+                'value': {
+                    'type': 'boolean',
+                    'load_only': 'true'
+                }
+            }
+            required = ['id']
+
+        user = {
+            'id': 1,
+            'name': 'test',
+            'value': False
+        }
+        assert SchemaRequiredList(**user)
+
+        bad_user = {
+            'name': 'test',
+            'value': False
+        }
+
+        with pytest.raises(ValueError):
+            SchemaRequiredList(**bad_user)
+
+    def test_schema_valid_when_required_is_set(self):
+        class SchemaRequiredSet(Schema):
+            type = 'object'
+            properties = {
+                'id': {
+                    'type': 'int'
+                },
+                'name': {
+                    'type': 'string',
+                },
+                'value': {
+                    'type': 'boolean',
+                    'load_only': 'true'
+                }
+            }
+            required = {'id'}
+
+        user = {
+            'id': 1,
+            'name': 'test',
+            'value': False
+        }
+        assert SchemaRequiredSet(**user)
+
+        bad_user = {
+            'name': 'test',
+            'value': False
+        }
+
+        with pytest.raises(ValueError):
+            SchemaRequiredSet(**bad_user)
+
+    def test_schema_valid_when_required_is_tuple(self):
+        class SchemaRequiredTuple(Schema):
+            type = 'object'
+            properties = {
+                'id': {
+                    'type': 'int'
+                },
+                'name': {
+                    'type': 'string',
+                },
+                'value': {
+                    'type': 'boolean',
+                    'load_only': 'true'
+                }
+            }
+            required = 'id',
+
+        user = {
+            'id': 1,
+            'name': 'test',
+            'value': False
+        }
+        assert SchemaRequiredTuple(**user)
+
+        bad_user = {
+            'name': 'test',
+            'value': False
+        }
+
+        with pytest.raises(ValueError):
+            SchemaRequiredTuple(**bad_user)
+
+    def test_schema_raise_error_when_required_is_bad_type(self):
+        with pytest.raises(TypeError):
+            class SchemaRequiredBadType(Schema):
+                type = 'object'
+                properties = {
+                    'id': {
+                        'type': 'int'
+                    },
+                    'name': {
+                        'type': 'string',
+                    },
+                    'value': {
+                        'type': 'boolean',
+                        'load_only': 'true'
+                    }
+                }
+                required = 'id'
+
+    def test_sub_schema_raise_error_when_change_type_of_existing_property(self, super_schema):
+        with pytest.raises(TypeError):
+            class SchemaPropertyTypeChange(super_schema):
+                properties = {
+                    'id': {'type': 'string'},
+                    'super_attribute': {'type': 'integer'}
+                }
+
+    def test_schema_array_example(self):
+        class SchemaArrayExample(Schema):
+            type = 'array'
+            items = {'type': 'string'}
+
+        assert SchemaArrayExample.example() == ['string']
+
+        class SchemaWillUseSchemaArray(Schema):
+            type = 'object'
+            properties = {'array': SchemaArrayExample}
+
+        assert json.loads(json.dumps(SchemaWillUseSchemaArray.example(), cls=ExampleEncoder)) == {
+            'array': ['string']
+        }
+
+        class SchemaWithArrayProperties(Schema):
+            type = 'object'
+            properties = {
+                'array': {
+                    'type': 'array',
+                    'items': {
+                        'type': 'string'
+                    }
+                }
+            }
+
+        assert json.loads(json.dumps(SchemaWithArrayProperties.example(), cls=ExampleEncoder)) == {
+            'array': ['string']
+        }
+
+    def test_schema_array_example_items_missing(self):
+        with pytest.raises(TypeError):
+            class SchemaArrayExampleItemsMissing(Schema):
+                type = 'array'
+
+    def test_sub_schema(self):
+        class AnEmailModel(Schema):
+            type = 'string'
+            # format = 'email'
+
+        class AnEmailOtherModel(Schema):
+            type = 'string'
+            format = 'email'
+
+        class NewOtherSchema(Schema):
+            type = 'object'
+            properties = {
+                'id': {
+                    'type': 'integer'
+                },
+                'name': {
+                    'type': 'string',
+                },
+                'value': {
+                    'type': 'boolean',
+                    'load_only': 'true'
+                },
+                'email': AnEmailModel
+            }
+
+        class NewOtherSubSchema(NewOtherSchema):
+            properties = {
+                'value': {
+                    'type': 'boolean',
+                    'load_only': 'false',
+                    'dump_only': 'true',
+                    'nullable': 'true'
+                },
+                'password': {
+                    'type': 'string'
+                },
+                'email': AnEmailOtherModel
+            }
+
+        properties = {
+            'id': {
+                'type': 'integer',
+            },
+            'name': {
+                'type': 'string',
+            },
+            'value': {
+                'type': 'boolean',
+                'load_only': 'true'
+            },
+            'email': AnEmailModel
+        }
+
+        new_properties = {
+            'id': {
+                'type': 'integer'
+            },
+            'name': {
+                'type': 'string',
+            },
+            'value': {
+                'type': 'boolean',
+                'load_only': 'false',
+                'nullable': 'true',
+                'dump_only': 'true',
+            },
+            'password': {
+                'type': 'string'
+            },
+            'email': {
+                'type': 'string',
+                'format': 'email'
+            }
+        }
+
+        assert NewOtherSchema.properties == properties
+
+        assert NewOtherSubSchema.properties == new_properties
+
+        class AnUserModel(Schema):
+            type = 'object'
+            properties = {
+                'id': {
+                    'type': 'integer'
+                },
+                'name': {
+                    'type': 'string'
+                },
+                'email': {
+                    'type': 'string'
+                }
+            }
+
+        class AMoreSeriousUserModel(Schema):
+            type = 'object'
+            properties = {
+                'id': {
+                    'type': 'integer'
+                },
+                'name': {
+                    'type': 'string'
+                },
+                'email': AnEmailOtherModel
+            }
+
+        class SuperClassThatUseUserModel(Schema):
+            type = 'object'
+            properties = {
+                'id': {
+                    'type': 'integer'
+                },
+                'user': AnUserModel
+            }
+
+        class SubClassThatUseUserModel(SuperClassThatUseUserModel):
+            properties = {
+                'user': AMoreSeriousUserModel
+            }
+
+        properties = {
+            'id': {
+                'type': 'integer',
+            },
+            'user': AnUserModel
+        }
+
+        new_properties = {
+            'id': {
+                'type': 'integer'
+            },
+            'user': {
+                'id': {
+                    'type': 'integer'
+                },
+                'name': {
+                    'type': 'string'
+                },
+                'email': {
+                    'type': 'string',
+                    'format': 'email'
+                }
+            }
+        }
+
+        assert SuperClassThatUseUserModel.properties == properties
+        assert SubClassThatUseUserModel.properties == new_properties